requirements.txt
PyDirectInput~=1.0.4
opencv-python~=4.5.1.48
Pillow>=8.3.2
numpy~=1.20.2
keyboard~=0.13.5